Add JUnit tests for ArticleRepository Refactoring entities

+++ src/main/java/com/siicanada/article/repository/ArticleRepository.java
@@ -3,6 +3,6 @@ package com.siicanada.article.repository;
 import com.siicanada.article.repository.entity.ArticleEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ArticleRepository extends JpaRepository<ArticleEntity, Long> {
+public interface ArticleRepository extends JpaRepository<ArticleEntity, Integer> {
 
 }
